KLUDGE: DON'T MERGE THIS COMMIT; IT WILL BE REBASED
This commit contains miscellaneous kludges to get
the code to run from a proc macro context.

The original code did unsafe things to access a private
compiler API. In order for this to work it made the assumption
that the compiler wasn't running in the context of a proc macro,
but I want to run the code in the context of a proc macro.

Fortunately the broken code is just for getting proper source map
line number information. The asserts are disabled and I'm not using
the line information.